requirements.txt
PyJWT==2.8.0
requests==2.31.0
PyMySQL==1.1.0
cloud-sql-python-connector[pymysql]==1.18.0
python-dotenv==1.0.0 
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Pillow